# Conflicts:
#	composer.lock
#	resources/views/template/initialScripts.blade.php

+++ resources/views/crud/rol/index.blade.php
@@ -1,4 +1,4 @@
-@extends('template.main')
+@extends('template.globalCard')
 
 
 @section('headerCard')
@@ -7,16 +7,15 @@
 
         <div class="tools">
             <div class="btn-group">
-                <a href="{{ route('rol.create') }}" type="button" class="btn btn-raised ink-reaction btn-primary btn-block" data-toggle="tooltip" data-original-title="New Role">New</a>
+                <a href="{{ route('admin.rol.create') }}" type="button" class="btn btn-raised ink-reaction btn-primary btn-block" data-toggle="tooltip" data-original-title="New Role">New</a>
             </div>
         </div>
     </div>
 @endsection
 
+@section('bodyCard')
 
-@section('content')
-
-    <div>
+    <div id="role">
         <table class="table table-hover">
             <thead>
                 <tr>
@@ -33,64 +32,21 @@
                         <td>{{$rol->name }}</td>
                         <td>{{$rol->description}}</td>
                         <td>
-                            <button type="button" class="btn btn-icon-toggle"  data-toggle="modal" data-target="#modalShow"data-original-title="Show {{$rol->name}}"><i class="fa fa-eye"></i></button>
-                            <a href="{{ route('rol.edit',$rol->id) }}" type="button" class="btn btn-icon-toggle"  data-toggle="tooltip" data-placement="top" data-original-title="Edit {{$rol->name}}"><i class="fa fa-pencil"></i></a>
-                            <button type="button" class="btn btn-icon-toggle" data-toggle="modal" data-target="#modalDelete" data-original-title="Delete {{$rol->name}}"><i class="fa fa-trash-o"></i></button>
+                            {{ Form::open(['route'=> ['admin.rol.destroy', $rol->id],'method'=> 'DELETE','onsubmit' => 'return confirm("are you sure ?")']) }}
+                                <a href="{{ route('admin.rol.show',$rol->id) }}"  class="btn btn-icon-toggle"  data-toggle="tooltip" data-original-title="Show {{$rol->name}}"><i class="fa fa-eye"></i></a>
+                                <a href="{{ route('admin.rol.edit',$rol->id) }}"  class="btn btn-icon-toggle"  data-toggle="tooltip" data-original-title="Edit {{$rol->name}}"><i class="fa fa-pencil"></i></a>
+                                <button type="submit" class="btn btn-icon-toggle" data-toggle="tooltip" data-original-title="Delete {{$rol->name}}"><i class="fa fa-trash-o"></i></button>
+                            {!! Form::close() !!}
                         </td>
                     </tr>
                 @endforeach
             </tbody>
 
         </table>
+        {!! $role->render() !!}
     </div>
+@endsection
 
-
-    <!-- Modal Show-->
-    <div id="modalShow" class="modal fade" role="dialog">
-        <div class="modal-dialog">
-
-            <!-- Modal content-->
-            <div class="modal-content">
-                <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal">&times;</button>
-                    <h4 class="modal-title">Role </h4>
-                </div>
-                <div class="modal-body">
-                    <p><strong>{{$rol->name}} :</strong>{{$rol->description}}</p>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                </div>
-            </div>
-
-        </div>
-    </div>
-
-    <!-- Modal Show-->
-    <div id="modalDelete" class="modal fade" role="dialog">
-        <div class="modal-dialog">
-
-            <!-- Modal content-->
-            <div class="modal-content">
-                <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal">&times;</button>
-                    <h4 class="modal-title">Do you want to delete the role? </h4>
-                </div>
-                <div class="modal-body">
-                    <p><strong>{{$rol->name}} :</strong>{{$rol->description}}</p>
-                </div>
-                <div class="modal-footer">
-
-                    {!! Form::open(['route'=> ['rol.destroy',$rol->id],'method'=> 'DELETE']) !!}
-                        <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
-                        <button type="submit" class="btn btn-danger">Delete</button>
-                    {!! Form::close() !!}
-
-                </div>
-            </div>
-
-        </div>
-    </div>
-
+@section('javascript')
 
 @endsection
